Fix linter errors & move (2)score functions to helper file
